+++ app/src/main/java/com/minicore/cartio/features/shopping/data/LocalShoppingListItemDataSource.kt
@@ -17,9 +17,9 @@ class LocalShoppingListItemDataSource @Inject constructor(
         dao.findActiveByProduct(listId, productId)
 
     override suspend fun insert(listId: Long, productId: Long): Long =
-        dao.insert(ShoppingListItemEntity(shoppingListId = listId, productId = productId, quantity = 1f))
+        dao.insert(ShoppingListItemEntity(shoppingListId = listId, productId = productId, quantity = 1))
 
-    override suspend fun updateQuantity(id: Long, quantity: Float) = dao.updateQuantity(id, quantity)
+    override suspend fun updateQuantity(id: Long, quantity: Int) = dao.updateQuantity(id, quantity)
 
     override suspend fun updateChecked(id: Long, checked: Boolean) = dao.updateChecked(id, checked)
